Lambda-Webhook/requirements-dev.txt
# Development Dependencies
# Copyright 2025 Loopper-AI

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto>=5.0.0  # AWS service mocking

# Type checking
mypy>=1.8.0
boto3-stubs[essential]>=1.34.0

# Linting and formatting
ruff>=0.1.0
black>=24.0.0

# Code quality
bandit>=1.7.0  # Security linting
